Refactored utils.js
